requirements.txt
PyInstaller
sqlalchemy~=2.0.38
pyyaml~=6.0.2
uvicorn~=0.32.1
fastapi~=0.112.2
pymysql~=1.0.2
pydantic~=2.10.3
pandas~=2.2.3
openpyxl~=3.1.5
python-jose[cryptography]~=3.3.0
bcrypt~=3.2.0
pyodbc~=5.2.0
python-multipart~=0.0.6
jose~=1.0.0
apscheduler~=3.11.0
pytz~=2024.1
xlsxwriter~=3.2.3
starlette~=0.38.2
paramiko~=3.5.0